refactor(apple): move `generate_*` functions to separate files
