+++ ios/test_app.rb
@@ -2,6 +2,8 @@ require('cfpropertylist')
 require('json')
 require('pathname')
 
+require_relative('assets_catalog')
+require_relative('info_plist')
 require_relative('pod_helpers')
 require_relative('privacy_manifest')
 require_relative('xcode')
@@ -44,10 +46,6 @@ def package_version(package_path)
   Gem::Version.new(package_json['version'])
 end
 
-def project_path(file, target_platform)
-  File.expand_path(file, File.join(__dir__, '..', target_platform.to_s))
-end
-
 def react_native_path(project_root, target_platform)
   @react_native_path ||= {}
 
@@ -71,86 +69,6 @@ def target_product_type(target)
   target.product_type if target.respond_to?(:product_type)
 end
 
-def generate_assets_catalog!(project_root, target_platform, destination)
-  xcassets_src = project_path('ReactTestApp/Assets.xcassets', target_platform)
-  xcassets_dst = File.join(destination, File.basename(xcassets_src))
-  FileUtils.rm_rf(xcassets_dst)
-  FileUtils.cp_r(xcassets_src, destination)
-
-  icons = platform_config('icons', project_root, target_platform)
-  return if icons.nil? || icons['primaryIcon'].nil?
-
-  template = JSON.parse(File.read(File.join(xcassets_src, 'AppIcon.appiconset', 'Contents.json')))
-  app_manifest_dir = File.dirname(find_file('app.json', project_root))
-
-  app_icons = (icons['alternateIcons'] || {}).merge({ 'AppIcon' => icons['primaryIcon'] })
-  app_icons.each do |icon_set_name, app_icon|
-    app_icon_set = File.join(xcassets_dst, "#{icon_set_name}.appiconset")
-    FileUtils.mkdir_p(app_icon_set)
-
-    icon = File.join(app_manifest_dir, app_icon['filename'])
-    extname = File.extname(icon)
-    basename = File.basename(icon, extname)
-
-    images = []
-
-    template['images'].each do |image|
-      scale, size = image.values_at('scale', 'size')
-      width, height = size.split('x')
-      filename = "#{basename}-#{height}@#{scale}#{extname}"
-      images << { 'filename' => filename }.merge(image)
-      fork do
-        output = File.join(app_icon_set, filename)
-        scale = scale.split('x')[0].to_f
-        height = height.to_f * scale
-        width = width.to_f * scale
-        `sips --resampleHeightWidth #{height} #{width} --out "#{output}" "#{icon}"`
-      end
-    end
-
-    File.write(File.join(app_icon_set, 'Contents.json'),
-               JSON.pretty_generate({ 'images' => images, 'info' => template['info'] }))
-  end
-end
-
-def generate_info_plist!(project_root, target_platform, destination)
-  manifest = app_manifest(project_root)
-  return if manifest.nil?
-
-  infoplist_src = project_path('ReactTestApp/Info.plist', target_platform)
-  infoplist_dst = File.join(destination, File.basename(infoplist_src))
-
-  plist = CFPropertyList::List.new(file: infoplist_src)
-  info = CFPropertyList.native_types(plist.value)
-
-  # Register fonts
-  font_files = ['.otf', '.ttf']
-  fonts = []
-  resources = resolve_resources(manifest, target_platform)
-  resources&.each do |filename|
-    fonts << File.basename(filename) if font_files.include?(File.extname(filename))
-  end
-  unless fonts.empty?
-    # https://developer.apple.com/documentation/bundleresources/information_property_list/atsapplicationfontspath
-    info['ATSApplicationFontsPath'] = '.' if target_platform == :macos
-    # https://developer.apple.com/documentation/uikit/text_display_and_fonts/adding_a_custom_font_to_your_app
-    info['UIAppFonts'] = fonts unless target_platform == :macos
-  end
-
-  if target_platform == :macos
-    config = manifest[target_platform.to_s]
-    unless config.nil?
-      category = config['applicationCategoryType']
-      info['LSApplicationCategoryType'] = category unless category.nil?
-      copyright = config['humanReadableCopyright']
-      info['NSHumanReadableCopyright'] = copyright unless copyright.nil?
-    end
-  end
-
-  plist.value = CFPropertyList.guess(info)
-  plist.save(infoplist_dst, CFPropertyList::List::FORMAT_XML, { :formatted => true })
-end
-
 def react_native_pods(version)
   if version.zero? || version >= v(0, 71, 0)
     'use_react_native-0.71'
@@ -163,13 +81,6 @@ def react_native_pods(version)
   else
     raise "Unsupported React Native version: #{version}"
   end
-end
-
-def resolve_resources(manifest, target_platform)
-  resources = manifest['resources']
-  return if !resources || resources.empty?
-
-  resources.instance_of?(Array) ? resources : resources[target_platform.to_s]
 end
 
 def validate_resources(resources, app_dir)
